feat: refactor IssueNotification structure and update imports across repositories
