+++ src/domain/entities/IssueNotification.ts
@@ -1,0 +1,12 @@
+export type IssueNotification = {
+    id: string;
+    subject: string;
+    messages: {
+        id: string;
+        name: string;
+        sender: {
+            id: string;
+            displayName: string;
+        };
+    }[];
+};
